Delete method updated in AuthController

+++ src/main/java/com/devb/usermanagement/entity/auth/AuthService.java
@@ -17,42 +17,39 @@ import com.devb.usermanagement.service.exception.ObjectNotFoundException;
 
 @Service
 public class AuthService {
-	
+
 	private final UserAppRepository userAppRepository;
 	private final PasswordEncoder passwordEncoder;
 	private final AuthenticationManager authenticationManager;
 	private final JwtService jwtService;
-	
-	
-	
+
 	public AuthService(UserAppRepository userAppRepository, PasswordEncoder passwordEncoder,
 			AuthenticationManager authenticationManager, JwtService jwtService) {
-		super();
+
 		this.userAppRepository = userAppRepository;
 		this.passwordEncoder = passwordEncoder;
 		this.authenticationManager = authenticationManager;
 		this.jwtService = jwtService;
 	}
 
-	public boolean checkUser(UserRegistrationRequest userRegistrationRequest){
+	public boolean checkUser(UserRegistrationRequest userRegistrationRequest) {
 		Optional<UserApp> userApp = userAppRepository.findByEmail(userRegistrationRequest.getEmail());
 		Optional<UserApp> useOptional = userAppRepository.findByPassw(userRegistrationRequest.getPassword());
-		
-		if(useOptional.isPresent() || userApp.isPresent()) {
+
+		if (useOptional.isPresent() || userApp.isPresent()) {
 			throw new DataIntegratyViolationException("Password or Email already exist");
 		}
 		return true;
 	}
-	
-	public UserApp checkEmail(String email){
+
+	public UserApp checkEmail(String email) {
 		Optional<UserApp> userApp = userAppRepository.findByEmail(email);
-		
-		return userApp.orElseThrow(
-				()-> new DataIntegratyViolationException("Email already exist"));
+
+		return userApp.orElseThrow(() -> new DataIntegratyViolationException("Email already exist"));
 	}
-	
+
 	public UserResponse save(UserRegistrationRequest userRegistrationRequest) {
-		
+
 		checkUser(userRegistrationRequest);
 		UserApp uApp = new UserApp();
 		uApp.setEmail(userRegistrationRequest.getEmail());
@@ -61,44 +58,50 @@ public class AuthService {
 		uApp.setPassw(passwordEncoder.encode(userRegistrationRequest.getPassword()));
 		uApp.setRole(Role.USER);
 		UserApp respApp = userAppRepository.save(uApp);
-		return new UserResponse(respApp.getId(),respApp.getFirstName()+ " "+respApp.getLastName(),
-				respApp.getEmail(),respApp.getRole().name());
+		return new UserResponse(respApp.getId(), respApp.getFirstName() + " " + respApp.getLastName(),
+				respApp.getEmail(), respApp.getRole().name());
 	}
-	
-	public LoginResponse login (UserAuthRequest userAuthRequest) {
-		 Authentication authentication = authenticationManager.authenticate(
-				 new UsernamePasswordAuthenticationToken(userAuthRequest.getEmail(), userAuthRequest.getPassword()));
-		 if(authentication.isAuthenticated()) {
-			 return new LoginResponse(authentication.getName(), jwtService.generateToken(authentication.getName())) ;
-		 }else {
-			 throw new DataIntegratyViolationException("Credentials invalid");
-		 }
+
+	public LoginResponse login(UserAuthRequest userAuthRequest) {
+		Authentication authentication = authenticationManager.authenticate(
+				new UsernamePasswordAuthenticationToken(userAuthRequest.getEmail(), userAuthRequest.getPassword()));
+
+		if (authentication.isAuthenticated()) {
+			return new LoginResponse(authentication.getName(), jwtService.generateToken(authentication.getName()));
+		} else {
+			throw new DataIntegratyViolationException("Credentials invalid");
+		}
 	}
-	public List<UserResponse> getAll(){
+
+	public List<UserResponse> getAll() {
 		List<UserApp> list = userAppRepository.findAll();
 		List<UserResponse> listResponse = new ArrayList<>();
 		list.forEach(resp -> {
-			listResponse.add(new UserResponse(resp.getId(),resp.getFirstName()+ " "+resp.getLastName(),resp.getEmail(),
-					resp.getRole().name()));
+			listResponse.add(new UserResponse(resp.getId(), resp.getFirstName() + " " + resp.getLastName(),
+					resp.getEmail(), resp.getRole().name()));
 		});
 		return listResponse;
 	}
-	
+
 	public UserApp findById(Integer id) {
 		Optional<UserApp> userOptional = userAppRepository.findById(id);
-		return userOptional.orElseThrow(()->
-			new ObjectNotFoundException("USER not FOUND"));
+		return userOptional.orElseThrow(() -> new ObjectNotFoundException("USER not FOUND"));
 	}
-	
+
 	public UserResponse update(Integer id, UserRegistrationRequest userRegistrationRequest) {
-		
+
 		Optional<UserApp> userApp = userAppRepository.findById(id);
 		return null;
 	}
-	public void deleteUser(Integer id) {
-		
+
+	public void deleteUser(Integer id, UserAuthRequest userAuthRequest) {
+
 		UserApp usOptional = findById(id);
+		if((usOptional.getEmail()
+				.equalsIgnoreCase(userAuthRequest.getEmail()) && (usOptional.getPassw()
+						.equalsIgnoreCase(userAuthRequest.getPassword())))) {
 		userAppRepository.delete(usOptional);
+		}
 	}
 
 }
